covid_tracker/requirements.txt
Jinja2==2.10.1
psycopg2==2.9.3
python-dotenv==1.0.0
requests==2.22.0
